feat: add iattention backend, refactor plugin converter, and address review comments
- Add --backend iattention using TRT native IAttention layer (PR #4104)
- Move plugin converter to plugin_converter.py (per review)
- Fix compile_plugin_model: drop enabled_precisions with use_explicit_typing, use Debugger
- Update README: document 3 backends, add MAX_JOBS=8, simplify performance section

Note: iattention backend currently falls back to manual matmul+softmax
decomposition because HuggingFace passes attn_bias (causal mask tensor) to
_scaled_dot_product_efficient_attention, causing the converter to bypass
TRT's native IAttention layer. See #4129 for tracking.
