+++ tools/llm/plugin_utils.py
@@ -8,6 +8,7 @@ with various LLM architectures (Qwen, Llama, etc.).
 import ctypes
 import inspect
 import os
+from contextlib import nullcontext
 from typing import Any, Callable, Dict, List, Optional, Tuple, Type
 
 import numpy as np
@@ -15,11 +16,6 @@ import tensorrt as trt
 import torch
 import torch.nn as nn
 import torch_tensorrt
-from torch_tensorrt.dynamo.conversion import (
-    ConversionContext,
-    dynamo_tensorrt_converter,
-)
-from torch_tensorrt.dynamo.conversion.converter_utils import get_trt_tensor
 
 # Default plugin path - can be overridden
 # Built from: https://github.com/chohk88/TensorRT-Edge-LLM/tree/feature/torch-tensorrt-python-runtime
@@ -184,88 +180,9 @@ if not (
 ):
     _register_plugin_op_impl()
 
-
-# -----------------------------------------------------------------------------
-# TensorRT Converter
-# -----------------------------------------------------------------------------
-
-
-@dynamo_tensorrt_converter(
-    torch.ops.tensorrt_edge_llm.xqa_attn.default, supports_dynamic_shapes=True
-)
-def convert_attn(ctx: ConversionContext, target, args, kwargs, name):
-    """
-    Convert tensorrt_edge_llm::xqa_attn op to TensorRT AttentionPlugin.
-
-    TensorRT-Edge-LLM (0.4.0) plugin requires 5 inputs:
-    - qkv, kv, ctx_len, rope, kv_cache_start_idx
-
-    Plugin fields:
-    - num_q_heads, num_kv_heads, head_size, enable_tree_attention, enable_delta_kv_output
-    """
-    # args: qkv, kv, ctx_len, rope, kv_cache_start_idx, nq, nkv, d
-    qkv, kv, ctx_len, rope, kv_cache_start_idx, nq, nkv, d = args[:8]
-
-    creator = trt.get_plugin_registry().get_plugin_creator("AttentionPlugin", "1", "")
-    if creator is None:
-        raise RuntimeError("AttentionPlugin not found in TensorRT plugin registry!")
-
-    # Get config from global settings
-    config = get_plugin_config()
-    if config:
-        nq_val = config["num_attention_heads"]
-        nkv_val = config["num_key_value_heads"]
-        d_val = config["head_dim"]
-    else:
-        # Fallback to values from args (may not work correctly)
-        nq_val = nq if isinstance(nq, int) else 14
-        nkv_val = nkv if isinstance(nkv, int) else 2
-        d_val = d if isinstance(d, int) else 64
-
-    # Plugin fields for TensorRT-Edge-LLM AttentionPlugin
-    # Required: num_q_heads, num_kv_heads, head_size, enable_tree_attention
-    # enable_delta_kv_output=1 enables delta KV output for Python/torch_tensorrt compatibility
-    field_list = [
-        trt.PluginField(
-            field_name, np.array([field_val], dtype=np.int32), trt.PluginFieldType.INT32
-        )
-        for field_name, field_val in [
-            ("num_q_heads", nq_val),
-            ("num_kv_heads", nkv_val),
-            ("head_size", d_val),
-            ("enable_tree_attention", 0),
-            ("enable_delta_kv_output", 1),
-        ]
-    ]
-
-    fields = trt.PluginFieldCollection(field_list)
-    plugin = creator.create_plugin(name, fields)
-
-    # 5 inputs for release version: qkv, kv, ctx_len, rope, kv_cache_start_idx
-    inputs = [
-        (
-            get_trt_tensor(ctx, i, f"{name}_i{idx}")
-            if not isinstance(i, trt.ITensor)
-            else i
-        )
-        for idx, i in enumerate([qkv, kv, ctx_len, rope, kv_cache_start_idx])
-    ]
-
-    # Handle ctx_len shape if needed (squeeze if [B, 1] -> [B])
-    if len(inputs[2].shape) == 2 and inputs[2].shape[1] == 1:
-        shuffle_layer = ctx.net.add_shuffle(inputs[2])
-        shuffle_layer.reshape_dims = (inputs[2].shape[0],)
-        inputs[2] = shuffle_layer.get_output(0)
-
-    # Handle kv_cache_start_idx shape if needed (squeeze if [B, 1] -> [B])
-    if len(inputs[4].shape) == 2 and inputs[4].shape[1] == 1:
-        shuffle_layer = ctx.net.add_shuffle(inputs[4])
-        shuffle_layer.reshape_dims = (inputs[4].shape[0],)
-        inputs[4] = shuffle_layer.get_output(0)
-
-    layer = ctx.net.add_plugin_v2(inputs, plugin)
-    return layer.get_output(0), layer.get_output(1)
-
+# The converter for tensorrt_edge_llm::xqa_attn is defined in plugin_converter.py.
+# Import it here so that importing plugin_utils still registers the converter.
+from plugin_converter import convert_attn  # noqa: F401
 
 # -----------------------------------------------------------------------------
 # RoPE Cache Generation
@@ -691,17 +608,16 @@ def compile_plugin_model(
     )
 
     # Compile
-    trt_model = torch_tensorrt.dynamo.compile(
-        ep,
-        inputs=[dummy_input_ids, dummy_pos_ids, dummy_kvs, dummy_ctx_len],
-        enabled_precisions={torch.float32},
-        use_explicit_typing=True,
-        use_fp32_acc=True,
-        device=device,
-        disable_tf32=True,
-        min_block_size=1,
-        debug=debug,
-    )
+    with torch_tensorrt.dynamo.Debugger() if debug else nullcontext():
+        trt_model = torch_tensorrt.dynamo.compile(
+            ep,
+            inputs=[dummy_input_ids, dummy_pos_ids, dummy_kvs, dummy_ctx_len],
+            use_explicit_typing=True,
+            use_fp32_acc=True,
+            device=device,
+            disable_tf32=True,
+            min_block_size=1,
+        )
 
     return trt_model
